backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
google-generativeai==0.7.2
google-adk==1.0.0
firebase-admin==6.5.0
requests==2.32.3
python-multipart==0.0.9
pydantic==2.7.1
